requirements.txt
asgiref==3.5.2
astroid==2.12.13
certifi==2022.9.24
cffi==1.15.1
cfgv==3.3.1
charset-normalizer==2.1.1
cloudinary==1.30.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==38.0.3
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.6
Django==4.1.3
django-cors-headers==3.13.0
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
filelock==3.8.0
identify==2.5.9
idna==3.4
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.2
lazy-object-proxy==1.8.0
MarkupSafe==2.1.1
mccabe==0.7.0
nodeenv==1.7.0
oauthlib==3.2.2
platformdirs==2.5.4
pre-commit==2.20.0
psycopg2-binary==2.9.5
pycparser==2.21
PyJWT==2.6.0
pylint==2.15.6
python3-openid==3.2.0
pytz==2022.6
PyYAML==6.0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.3
toml==0.10.2
tomli==2.0.1
tomlkit==0.11.6
uritemplate==4.1.1
urllib3==1.26.12
virtualenv==20.16.7
wrapt==1.14.1
